# src/main/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.2.3-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `https://MY-REGISTRY-URL/apis/registry/v2` by default. Therefore you must prefix all API operation paths with `../apis/registry/v2` in this case. For example: `../apis/registry/v2/ids/globalIds/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
...
    },
    "paths": {
        "/ids/contentIds/{contentId}/": {
            "summary": "Access artifact content utilizing the unique content identifier for that content.",
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ArtifactContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getContentById",
                "summary": "Get artifact content by ID",
                "description": "Gets the content for an artifact version in the registry using the unique content\nidentifier for that content.  This content ID may be shared by multiple artifact\nversions in the case where the artifact versions are identical.\n\nThis operation may fail for one of the following reasons:\n\n* No content with this `contentId` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "contentId",
                    "description": "Global identifier for a single artifact content.",
                    "schema": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/ids/globalIds/{globalId}": {
            "summary": "Access artifact content utilizing an artifact version's globally unique identifier.",
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "dereference",
                        "description": "Allows the user to specify if the content should be dereferenced when being returned",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ArtifactContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getContentByGlobalId",
                "summary": "Get artifact by global ID",
                "description": "Gets the content for an artifact version in the registry using its globally unique\nidentifier.\n\nThis operation may fail for one of the following reasons:\n\n* No artifact version with this `globalId` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "globalId",
                    "description": "Global identifier for an artifact version.",
                    "schema": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/ids/contentHashes/{contentHash}/": {
            "summary": "Access artifact content utilizing the SHA-256 hash of the content.",
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ArtifactContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getContentByHash",
                "summary": "Get artifact content by SHA-256 hash",
                "description": "Gets the content for an artifact version in the registry using the \nSHA-256 hash of the content.  This content hash may be shared by multiple artifact\nversions in the case where the artifact versions have identical content.\n\nThis operation may fail for one of the following reasons:\n\n* No content with this `contentHash` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "contentHash",
                    "description": "SHA-256 content hash for a single artifact content.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/ids/contentHashes/{contentHash}/references": {
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "contentHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArtifactReference"
                                    }
                                }
                            }
                        },
                        "description": "A list containing all the references for the artifact with the given content hash."
                    }
                },
                "operationId": "referencesByContentHash",
                "summary": "List artifact references with given hash",
                "description": "Returns a list containing all the artifact references using the artifact content hash.\n\nThis operation may fail for one of the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            }
        },
        "/ids/contentIds/{contentId}/references": {
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "contentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArtifactReference"
                                    }
                                }
                            }
                        },
                        "description": "A list containing all the references for the artifact with the given content id."
                    }
                },
                "operationId": "referencesByContentId",
                "summary": "List artifact references with given content ID",
                "description": "Returns a list containing all the artifact references using the artifact content ID.\n\nThis operation may fail for one of the following reasons:\n\n* A server error occurred (HTTP error `500`)"
            }
        },
        "/ids/globalIds/{globalId}/references": {
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "globalId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArtifactReference"
                                    }
                                }
                            }
                        },
                        "description": "A list containing all the references for the artifact with the given global id."
                    }
                },
                "operationId": "referencesByGlobalId",
                "summary": "List artifact references with given global ID",
                "description": "Returns a list containing all the artifact references using the artifact global ID.\n\nThis operation may fail for one of the following reasons:\n\n* A server error occurred (HTTP error `500`)"
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Root Type for Error",
                "description": "All error responses, whether `4xx` or `5xx` will include one of these as the response\nbody.",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "The short error message.",
                        "type": "string"
                    },
                    "error_code": {
                        "format": "int32",
                        "description": "The server-side error code.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "Full details about the error.  This might contain a server stack trace, for example.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The error name - typically the classname of the exception thrown by the server.",
                        "type": "string"
                    }
                },
                "example": {
                    "error_code": 500,
                    "message": "An error occurred somewhere."
                }
            },
...
            "ArtifactReference": {
                "title": "Root Type for ArtifactReference",
                "description": "A reference to a different artifact. Typically used with artifact types that can have dependencies like Protobuf.",
                "required": [
                    "artifactId",
                    "groupId",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "groupId": {
                        "type": "string"
                    },
                    "artifactId": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "example": {
                    "groupId": "mygroup",
                    "artifactId": "13842090-2ce3-11ec-8d3d-0242ac130003",
                    "version": "2",
                    "name": "foo.bar.Open"
                }
            },
...
        },
        "responses": {
            "NotFound": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "NotFoundExample": {
                                "value": {
                                    "error_code": 404,
                                    "message": "No artifact with id 'Topic-1/Inbound' could be found."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can return a `404` error."
            },
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
...
            "ArtifactContent": {
                "content": {
                    "*/*": {
                        "schema": {

                        },
                        "examples": {
                            "OpenAPI": {
                                "value": {
                                    "openapi": "3.0.2",
                                    "info": {
                                        "title": "Empty API",
                                        "version": "1.0.0",
                                        "description": "An example API design using OpenAPI."
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "The content of one version of one artifact."
            }
        }
    },
...
}